added support for cryptocurrency BLOC

+++ src/gui/miner/spec.go
@@ -36,7 +36,7 @@ type Stats struct {
 	// HashrateHuman is the H/s, KH/s or MH/s representation of hashrate
 	HashrateHuman string `json:"hashrate_human"`
 	// CurrentDifficulty as set by the pool
-	CurrentDifficulty int `json:"current_difficulty"`
+	CurrentDifficulty int64 `json:"current_difficulty"`
 	// SharesGood is the good shares counter
 	SharesGood int `json:"shares_good"`
 	// SharesGood is the bad shares counter
